+++ scripts/addons/cam/ui/panels/area.py
@@ -18,7 +18,7 @@ class CAM_AREA_Panel(CAMButtonsPanel, Panel):
     bl_category = "CNC"
 
     bl_label = "[ Operation Area ]"
-    bl_idname = "WORLD_PT_CAM_OPERATION_AREA"
+    bl_idname = "FABEX_PT_CAM_OPERATION_AREA"
     panel_interface_level = 0
 
     def draw(self, context):
