+++ scripts/addons/cam/ui/panels/gcode.py
@@ -17,7 +17,7 @@ class CAM_GCODE_Panel(CAMButtonsPanel, Panel):
     bl_category = "CNC"
 
     bl_label = "[ Operation G-code ]"
-    bl_idname = "WORLD_PT_CAM_GCODE"
+    bl_idname = "FABEX_PT_CAM_GCODE"
     panel_interface_level = 1
 
     def draw(self, context):
