+++ scripts/addons/cam/ui/panels/interface.py
@@ -119,18 +119,18 @@ def update_layout(self, context):
     # Unregister all permanent panels
     try:
         unregister_classes = [
-            bpy.types.WORLD_PT_CAM_OPERATION_AREA,
-            bpy.types.WORLD_PT_CAM_CHAINS,
-            bpy.types.WORLD_PT_CAM_CUTTER,
-            bpy.types.WORLD_PT_CAM_FEEDRATE,
-            bpy.types.WORLD_PT_CAM_GCODE,
-            bpy.types.WORLD_PT_CAM_INFO,
-            bpy.types.WORLD_PT_CAM_MACHINE,
-            bpy.types.WORLD_PT_CAM_MATERIAL,
-            bpy.types.WORLD_PT_CAM_MOVEMENT,
-            bpy.types.WORLD_PT_CAM_OPERATION,
-            bpy.types.WORLD_PT_CAM_OPERATIONS,
-            bpy.types.WORLD_PT_CAM_OPTIMISATION,
+            bpy.types.FABEX_PT_CAM_OPERATION_AREA,
+            bpy.types.FABEX_PT_CAM_CHAINS,
+            bpy.types.FABEX_PT_CAM_CUTTER,
+            bpy.types.FABEX_PT_CAM_FEEDRATE,
+            bpy.types.FABEX_PT_CAM_GCODE,
+            bpy.types.FABEX_PT_CAM_INFO,
+            bpy.types.FABEX_PT_CAM_MACHINE,
+            bpy.types.FABEX_PT_CAM_MATERIAL,
+            bpy.types.FABEX_PT_CAM_MOVEMENT,
+            bpy.types.FABEX_PT_CAM_OPERATION,
+            bpy.types.FABEX_PT_CAM_OPERATIONS,
+            bpy.types.FABEX_PT_CAM_OPTIMISATION,
             bpy.types.VIEW3D_PT_tools_curvetools,
             bpy.types.VIEW3D_PT_tools_create,
         ]
