+++ scripts/addons/cam/ui/panels/movement.py
@@ -20,7 +20,7 @@ class CAM_MOVEMENT_Panel(CAMButtonsPanel, Panel):
     bl_category = "CNC"
 
     bl_label = "[ Movement ]"
-    bl_idname = "WORLD_PT_CAM_MOVEMENT"
+    bl_idname = "FABEX_PT_CAM_MOVEMENT"
     panel_interface_level = 0
 
     def draw(self, context):
