+++ scripts/addons/cam/ui/pie_menu/pie_chains.py
@@ -25,11 +25,11 @@ class VIEW3D_MT_PIE_Chains(Menu):
 
         # Left
         pie.operator("wm.call_panel", text="Operations", icon="MOD_ENVELOPE").name = (
-            "WORLD_PT_CAM_OPERATIONS"
+            "FABEX_PT_CAM_OPERATIONS"
         )
 
         # Right
-        pie.operator("wm.call_panel", text="Chains", icon="LINKED").name = "WORLD_PT_CAM_CHAINS"
+        pie.operator("wm.call_panel", text="Chains", icon="LINKED").name = "FABEX_PT_CAM_CHAINS"
 
         # Bottom
         row = pie.row()
